move bookings model to module folder
