+++ config/lighthouse.php
@@ -79,7 +79,7 @@ return [
     */
 
     'namespaces' => [
-        'models' => ['App', 'App\\Models', 'App\\Modules\\Endorsement\\Special'],
+        'models' => ['App', 'App\\Models', 'App\\Modules\\Endorsement\\Special', 'App\\Modules\\Bookings'],
         'queries' => 'App\\GraphQL\\Queries',
         'mutations' => 'App\\GraphQL\\Mutations',
         'subscriptions' => 'App\\GraphQL\\Subscriptions',
