+++ database/factories/BookingFactory.php
@@ -1,7 +1,8 @@
 <?php
 
 /** @var \Illuminate\Database\Eloquent\Factory $factory */
-use App\Booking;
+
+use App\Modules\Bookings\Booking;
 use Faker\Generator as Faker;
 
 $factory->define(Booking::class, function (Faker $faker) {
